Extract fetch-handler into a separate file
